+++ sys/dev/e1000/e1000_i210.c
@@ -716,6 +716,6 @@ s32 e1000_init_hw_i210(struct e1000_hw *hw)
 	/* Initialize identification LED */
 	mac->ops.id_led_init(hw);
 
-	ret_val = e1000_init_hw_82575(hw);
+	ret_val = e1000_init_hw_base(hw);
 	return ret_val;
 }
